Lazy load less only when needed
In order to not load `less` and thus `therubyracer` on production
systems, we utilize the same way sprockets uses for autoloading extra
dependencies.

- don't require `less` upfront
- autoload `less` and related code on demand
- access `Parser` via `[]` instead of the method to not confuse it with
  actual `Parser` class.
