+++ lib/less/rails.rb
@@ -3,7 +3,6 @@ module Less
   end
 end
 
-require 'less'
 require 'rails'
 require 'sprockets'
 begin
@@ -13,7 +12,6 @@ rescue LoadError
 end
 
 require 'less/rails/version'
-require 'less/rails/helpers'
 require 'less/rails/less_transformer'
 require 'less/rails/import_processor'
 require 'less/rails/railtie'
