+++ lib/less/rails/autoload.rb
@@ -1,0 +1,9 @@
+module Less
+  module Rails
+    module Autoload
+
+      autoload :Less, 'less/rails/autoload/less'
+
+    end
+  end
+end
