+++ lib/rails/generators/less/scaffold/scaffold_generator.rb
@@ -1,17 +1,17 @@
-require "less"
+require "less/rails/autoload"
 require "rails/generators/named_base"
 require "rails/generators/rails/scaffold/scaffold_generator"
 
 module Less
   module Generators
     class ScaffoldGenerator < ::Rails::Generators::NamedBase
-      
+
       def copy_stylesheet
         file = File.join ::Rails::Generators::ScaffoldGenerator.source_root, 'scaffold.css'
-        css = Less::Parser.new.parse(File.read(file)).to_css
+        css = Less::Rails::Autoload::Less::Parser.new.parse(File.read(file)).to_css
         create_file "app/assets/stylesheets/scaffolds.css.less", css
       end
-      
+
     end
   end
 end
